Add files via upload
This case study aims to identify patterns which indicate if a client has difficulty in paying their installments which may be used for taking actions such as denying the loan, reducing the amount of loan, lending (to risky applicants) at a higher interest rate, etc. This case study will ensure that the consumers capable of repaying the loan are not rejected. Identification of such applicants using EDA is the aim of this case study.

The company wants to understand the driving factors (or driver variables) behind loan default, i.e. the variables which are strong indicators of defaulter. The company can utilize this knowledge for its portfolio and risk assessment.